compiled

+++ src/miniprojet/ComparateurDeChaine.java
@@ -1,5 +1,5 @@
 package miniprojet;
 
 public interface ComparateurDeChaine extends ComparateurDeNom {
-    public double Comparer(Nom nom1,Nom nom2 );
+    public int Comparer(Nom nom1,Nom nom2 );
 }
